+++ src/admin/alavirien.js
@@ -37,7 +37,7 @@ const giveAlavirien = async (client, server, personality, userId) => {
   if (!guildMember.roles.cache.has(server.alavirienRoleId)) {
     //if doesn't have the role
     guildMember.roles.add(server.alavirienRoleId); //add role
-    removeAlavirien(userId, client.db); //remove from db
+    removeAlavirien(client.db, userId); //remove from db
 
     //send log
     const embed = setupEmbed("DARK_GREY", personality, guildMember.user, "tag"); //create log