+++ src/commands/polls/pollsSettings.js
@@ -11,7 +11,11 @@ import { COMMONS } from "../../commons.js";
 
 export const sendSettingsButtons = async (interaction) => {
   console.log("sendSettingsButtons");
-  await interaction.deferReply({ ephemeral: true });
+  try {
+    await interaction.deferReply({ ephemeral: true });
+  } catch (e) {
+    return console.log(e);
+  }
 
   //get personality
   const perso = PERSONALITY.getCommands().polls.settings;
@@ -56,7 +60,11 @@ export const sendSettingsButtons = async (interaction) => {
 
 export const disablePoll = async (interaction) => {
   console.log("disablePoll");
-  await interaction.deferUpdate();
+  try {
+    await interaction.deferUpdate();
+  } catch (e) {
+    return console.log(e);
+  }
 
   //get personality
   const perso = PERSONALITY.getCommands().polls.settings;
@@ -89,7 +97,11 @@ export const disablePoll = async (interaction) => {
 };
 
 export const removePollButtonAction = async (interaction) => {
-  await interaction.deferUpdate();
+  try {
+    await interaction.deferUpdate();
+  } catch (e) {
+    return console.log(e);
+  }
 
   const perso = PERSONALITY.getCommands().polls.settings.remove;
 
@@ -126,7 +138,11 @@ export const removePollButtonAction = async (interaction) => {
 };
 
 export const resetPollButtonAction = async (interaction) => {
-  await interaction.deferUpdate();
+  try {
+    await interaction.deferUpdate();
+  } catch (e) {
+    return console.log(e);
+  }
 
   //get data
   const personality = PERSONALITY.getCommands().polls;
@@ -160,7 +176,11 @@ export const resetPollButtonAction = async (interaction) => {
 
 export const updatePollButtonAction = async (interaction) => {
   console.log("updatePollButtonAction");
-  await interaction.deferUpdate();
+  try {
+    await interaction.deferUpdate();
+  } catch (e) {
+    return console.log(e);
+  }
 
   //get personality
   const perso = PERSONALITY.getCommands().polls.settings.update;